Update index.html
Put together a function that builds the url and a function that sends the request. The request looks like it goes through all right on the console, but parsing something out of the response still needs work.
Had to change the date function because it wouldn't plug in as it was before.